Update template_srv getVariables to return new TypedVariableModel
