+++ public/app/features/plugins/datasource_srv.ts
@@ -24,8 +24,6 @@ import {
   instanceSettings as expressionInstanceSettings,
 } from 'app/features/expressions/ExpressionDatasource';
 
-import { isDataSource } from '../variables/guard';
-
 import { importDataSourcePlugin } from './plugin_loader';
 
 export class DatasourceSrv implements DataSourceService {
@@ -246,7 +244,7 @@ export class DatasourceSrv implements DataSourceService {
 
     if (filters.variables) {
       for (const variable of this.templateSrv.getVariables()) {
-        if (!isDataSource(variable) || variable.multi || variable.includeAll) {
+        if (variable.type !== 'datasource' || variable.multi || variable.includeAll) {
           continue;
         }
         const dsName = variable.current.value === 'default' ? this.defaultName : variable.current.value;
