+++ public/app/features/templating/template_srv.ts
@@ -1,13 +1,19 @@
 import { escape, isString, property } from 'lodash';
 
-import { deprecationWarning, ScopedVars, TimeRange } from '@grafana/data';
+import {
+  deprecationWarning,
+  ScopedVars,
+  TimeRange,
+  AdHocVariableFilter,
+  AdHocVariableModel,
+  TypedVariableModel,
+} from '@grafana/data';
 import { getDataSourceSrv, setTemplateSrv, TemplateSrv as BaseTemplateSrv } from '@grafana/runtime';
 
 import { variableAdapters } from '../variables/adapters';
 import { ALL_VARIABLE_TEXT, ALL_VARIABLE_VALUE } from '../variables/constants';
 import { isAdHoc } from '../variables/guard';
 import { getFilteredVariables, getVariables, getVariableWithName } from '../variables/state/selectors';
-import { AdHocVariableFilter, AdHocVariableModel, VariableModel } from '../variables/types';
 import { variableRegex } from '../variables/utils';
 
 import { FormatOptions, formatRegistry, FormatRegistryID } from './formatRegistry';
@@ -56,8 +62,8 @@ export class TemplateSrv implements BaseTemplateSrv {
     return this.getVariables();
   }
 
-  getVariables(): VariableModel[] {
-    return this.dependencies.getVariables();
+  getVariables(): TypedVariableModel[] {
+    return this.dependencies.getVariables() as TypedVariableModel[];
   }
 
   updateIndex() {
