+++ public/app/features/variables/types.ts
@@ -4,10 +4,56 @@ import {
   BusEventWithPayload,
   DataQuery,
   DataSourceJsonData,
-  DataSourceRef,
   LoadingState,
   QueryEditorProps,
-  VariableModel as BaseVariableModel,
+  BaseVariableModel,
+  VariableHide,
+} from '@grafana/data';
+export {
+  /** @deprecated Import from @grafana/data instead */
+  VariableRefresh,
+  /** @deprecated Import from @grafana/data instead */
+  VariableSort,
+  /** @deprecated Import from @grafana/data instead */
+  VariableHide,
+  /** @deprecated Import from @grafana/data instead */
+  AdHocVariableFilter,
+  /** @deprecated Import from @grafana/data instead */
+  AdHocVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  VariableOption,
+  /** @deprecated Import from @grafana/data instead */
+  IntervalVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  CustomVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  DataSourceVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  QueryVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  TextBoxVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  ConstantVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  VariableWithMultiSupport,
+  /** @deprecated Import from @grafana/data instead */
+  VariableWithOptions,
+  /** @deprecated Import from @grafana/data instead */
+  DashboardProps,
+  /** @deprecated Import from @grafana/data instead */
+  DashboardVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  OrgProps,
+  /** @deprecated Import from @grafana/data instead */
+  OrgVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  UserProps,
+  /** @deprecated Import from @grafana/data instead */
+  UserVariableModel,
+  /** @deprecated Import from @grafana/data instead */
+  SystemVariable,
+  /** @deprecated Import from @grafana/data instead */
+  BaseVariableModel as VariableModel,
 } from '@grafana/data';
 import { TemplateSrv } from '@grafana/runtime';
 
@@ -19,155 +65,7 @@ export enum TransactionStatus {
   Completed = 'Completed',
 }
 
-export enum VariableRefresh {
-  never, // removed from the UI
-  onDashboardLoad,
-  onTimeRangeChanged,
-}
-
-export enum VariableHide {
-  dontHide,
-  hideLabel,
-  hideVariable,
-}
-
-export enum VariableSort {
-  disabled,
-  alphabeticalAsc,
-  alphabeticalDesc,
-  numericalAsc,
-  numericalDesc,
-  alphabeticalCaseInsensitiveAsc,
-  alphabeticalCaseInsensitiveDesc,
-}
-
-export interface VariableOption {
-  selected: boolean;
-  text: string | string[];
-  value: string | string[];
-  isNone?: boolean;
-}
-
-export interface AdHocVariableFilter {
-  key: string;
-  operator: string;
-  value: string;
-  condition: string;
-}
-
-export interface AdHocVariableModel extends VariableModel {
-  type: 'adhoc';
-  datasource: DataSourceRef | null;
-  filters: AdHocVariableFilter[];
-}
-
-export interface IntervalVariableModel extends VariableWithOptions {
-  type: 'interval';
-  auto: boolean;
-  auto_min: string;
-  auto_count: number;
-  refresh: VariableRefresh;
-}
-
-export interface CustomVariableModel extends VariableWithMultiSupport {
-  type: 'custom';
-}
-
-export interface DataSourceVariableModel extends VariableWithMultiSupport {
-  type: 'datasource';
-  regex: string;
-  refresh: VariableRefresh;
-}
-
-export interface QueryVariableModel extends VariableWithMultiSupport {
-  type: 'query';
-  datasource: DataSourceRef | null;
-  definition: string;
-  sort: VariableSort;
-  queryValue?: string;
-  query: any;
-  regex: string;
-  refresh: VariableRefresh;
-}
-
-export interface TextBoxVariableModel extends VariableWithOptions {
-  type: 'textbox';
-  originalQuery: string | null;
-}
-
-export interface ConstantVariableModel extends VariableWithOptions {
-  type: 'constant';
-}
-
-export interface VariableWithMultiSupport extends VariableWithOptions {
-  multi: boolean;
-  includeAll: boolean;
-  allValue?: string | null;
-}
-
-export interface VariableWithOptions extends VariableModel {
-  current: VariableOption;
-  options: VariableOption[];
-  query: string;
-}
-
-export interface DashboardProps {
-  name: string;
-  uid: string;
-  toString: () => string;
-}
-
-export interface DashboardVariableModel extends SystemVariable<DashboardProps> {}
-
-export interface OrgProps {
-  name: string;
-  id: number;
-  toString: () => string;
-}
-
-export interface OrgVariableModel extends SystemVariable<OrgProps> {}
-
-export interface UserProps {
-  login: string;
-  id: number;
-  email?: string;
-  toString: () => string;
-}
-
-export interface UserVariableModel extends SystemVariable<UserProps> {}
-
-export interface SystemVariable<TProps extends { toString: () => string }> extends VariableModel {
-  type: 'system';
-  current: { value: TProps };
-}
-
-export interface VariableModel extends BaseVariableModel {
-  name: string;
-  label?: string;
-  id: string;
-  rootStateKey: string | null;
-  global: boolean;
-  hide: VariableHide;
-  skipUrlSync: boolean;
-  index: number;
-  state: LoadingState;
-  error: any | null;
-  description: string | null;
-}
-
-export type TypedVariableModel =
-  | QueryVariableModel
-  | AdHocVariableModel
-  | ConstantVariableModel
-  | DataSourceVariableModel
-  | IntervalVariableModel
-  | TextBoxVariableModel
-  | CustomVariableModel
-  | UserVariableModel
-  | OrgVariableModel
-  | DashboardVariableModel;
-
-export const initialVariableModelState: VariableModel = {
+export const initialVariableModelState: BaseVariableModel = {
   id: NEW_VARIABLE_ID,
   rootStateKey: null,
   name: '',
